+++ init.php
@@ -16,7 +16,4 @@ define( 'DECALOG_VERSION', '1.10.0-dev0' );
 define( 'DECALOG_CODENAME', '"-"' );
 define( 'DECALOG_CRON_NAME', 'decalog_clean_database' );
 
-define( 'DECALOG_MINIMUM_WP_VERSION', '5.2' );
-define( 'DECALOG_MINIMUM_PHP_VERSION', '7.2' );
-
 define( 'DECALOG_CDN_AVAILABLE', true );