+++ includes/system/class-cache.php
@@ -511,7 +511,8 @@ class Cache {
 		$log      .= '   Hit count: ' . $analytics['hit']['count'] . '   Hit time: ' . round($analytics['hit']['time'] * 1000, 3) . 'ms   Hit size: ' . Conversion::data_shorten( (int) $analytics['hit']['size'] );
 		$log      .= '   Miss count: ' . $analytics['miss']['count'] . '   Miss time: ' . round($analytics['miss']['time'] * 1000, 3) . 'ms   Miss size: ' . Conversion::data_shorten( (int) $analytics['miss']['size'] );
 		if ( 0 !== (int) $analytics['hit']['count'] || 0 !== (int) $analytics['miss']['count'] ) {
-			Logger::debug( $log );
+			$logger = new Logger( 'plugin', DECALOG_PRODUCT_NAME, DECALOG_VERSION );
+			$logger->debug( $log );
 		}
 	}