+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased - Will be 1.10.0]
 ### Added
 - New listener for Wordfence plugin.
-- Compatibility with PerfOps.One suite.
+- Full integration with PerfOps.One suite.
 ### Changed
 - New menus (in the left admin bar) for accessing features: "PerfOps Records" and "PerfOps Settings".
 ### Fixed