+++ CHANGELOG.md
@@ -3,16 +3,18 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - Will be 1.10.0]
+## [1.10.0] - 2020-03-01
 ### Added
 - New listener for Wordfence plugin.
 - Full integration with PerfOps.One suite.
+- Compatibility with WordPress 5.4.
 ### Changed
 - New menus (in the left admin bar) for accessing features: "PerfOps Records" and "PerfOps Settings".
 ### Fixed
 - The shutdown action of [APCu Manager](https://wordpress.org/plugins/apcu-manager/) can cause a PHP notice in DecaLog.
 ### Removed
 - Compatibility with WordPress versions prior to 5.2.
+- Old menus entries, due to PerfOps integration.
 
 ## [1.9.1] - 2020-02-13
 ### Changed